+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Version 1.2.26
 
   * Add fetching of file names for events (unable to download)
+  * Add fetching of payment schedules - Osm::OnlinePayment::Schedule
 
 ## Version 1.2.25